refactor: use TryParseHex() instead of IsHex() and ParseHex()
TryParseHex() was introduced in #25227 and returns a std::nullopt if
the string is an invalid hex representation. This allows us to simplify
code that would first check if a string is valid hex with IsHex(),
and then iterate over the string again with ParseHex().

There is slight behaviour difference between these approaches, in that
IsHex() returns false for an empty string, whereas TryParseHex()
does not fail. For the updated instances, this does not make a difference.