scripted-diff: move files from util to common
Move util library files that are not used by the kernel library to the
common library. This is consistent with the library structure laid out
in doc/design/libraries.md.

Moving util/message.cpp to the common library also breaks an
undocumented dependency of the util library on the consensus library's
CPubKey::RecoverCompact symbol.

-BEGIN VERIFY SCRIPT-
FILES_TO_MOVE=(
    "asmap.cpp"
    "asmap.h"
    "bip32.cpp"
    "bip32.h"
    "bytevectorhash.cpp"
    "bytevectorhash.h"
    "error.cpp"
    "error.h"
    "fees.cpp"
    "fees.h"
    "message.cpp"
    "message.h"
    "readwritefile.cpp"
    "readwritefile.h"
    "sock.cpp"
    "sock.h"
    "spanparsing.cpp"
    "spanparsing.h"
)

for file in "${FILES_TO_MOVE[@]}"; do
    mv "src/util/$file" "src/common/$file"

    # If the file is a header file, update the include guard
    if [[ $file == *.h ]]; then
        sed -i 's/BITCOIN_UTIL_/BITCOIN_COMMON_/' "src/common/$file"
    fi

    # Update include statements, Makefile.am and other references for this file
    find src/ -type f \( -name "*.cpp" -o -name "*.h" \) -exec sed -i "s|util/$file|common/$file|" {} +
    sed -i "s|util/$file|common/$file|g" src/Makefile.am
done
-END VERIFY SCRIPT-