fixup: replace NeedsRateLimiting with Consume
By introducing the Status enum, we can let the callsite deal with
e.g. raising an excessive logging warning, and improve testability.
This way, `Consume` can take a `const str` parameter.

Also exposes a SuppressionsActive() method so the logger can use
that in a next commit to prefix [*] to logs whenenever suppressions
are active.